# Conflicts:
#	pa4/controllers/bookmarks.js

+++ pa4/controllers/bookmarks.js
@@ -1,23 +1,8 @@
-var _         = require('lodash');
-var db        = require('../database/db');
-var sql       = require('sql-query'), sqlQuery = sql.Query();
-var converter = require('../services/csvToJson');
+var _        = require('lodash');
+var db       = require('../database/db');
+var sql      = require('sql-query'), sqlQuery = sql.Query();
+var csv      = require('../services/csvToJson');
 var validUrl = require('valid-url'); //npm install valid-url
-var reportedError = null;
-
-function parseCSVFile(buffer, onNewRecord, handleError, done) {
-    var source = String.fromCharCode.apply(null, buffer);
-
-    converter.fromString(source, function (err, result) {
-        if (err)
-        {
-            handleError(err)
-        }
-
-        onNewRecord(result);
-        done();
-    });
-}
 
 module.exports.parseFile = function parseFile(req, res, next) {
     var buffer = req.file.buffer;
@@ -32,7 +17,7 @@ module.exports.parseFile = function parseFile(req, res, next) {
 
             insertBookmark(record[key], function (err) {
                 console.error(err);
-                throw err;
+                req.reportedError = err;
             }, function (el) {
                 console.log(`Successfully inserted ${JSON.stringify(el)}`);
             });
@@ -49,7 +34,7 @@ module.exports.parseFile = function parseFile(req, res, next) {
         renderIndex(req, res);
     }
 
-    parseCSVFile(buffer, onNewRecord, onError, done);
+    csv.parseCSVFile(buffer, onNewRecord, onError, done);
 
 };
 
@@ -57,57 +42,49 @@ var list = module.exports.list = function (req, res) {
     renderIndex(req, res);
 };
 
-
-
-
 function renderIndex(req, res, scopeCallBack) {
 
-    if (reportedError != null)
-    {
-        console.error(reportedError);
-        res.render('index', {error: err});
-        reportedError = null;
-        return;
-    }
-
-    var search = req.query['search'] ? db.escape(req.query.search): null;
     var sql;
-    //var sqlSelect = sqlQuery.select();
-    console.info('List request', req.query);
-    var folderId = req.query['folderId'] ? db.escape(req.query.folderId) : 1;
-    var sortBy   = req.query['sortBy'] ? (req.query.sortBy) : 'name';
+    var search   = req.query['search'] ? db.escape(req.query.search) : null;
+    var folderId = req.query['folderId'] ? db.escape(req.query.folderId) : req.session.folderId ? req.session.folderId : 1;
+    var sortBy   = req.query['sortBy'] ? req.query.sortBy : req.session.sortBy ? req.session.sortBy : 'name';
 
     req.session.folderId = folderId;
-    req.session.sortBy = sortBy;
+    req.session.sortBy   = sortBy;
 
-    if(search){
-        search = '% ' + req.query['search'] ? db.escape(req.query.search): null + ' %';
-        sql = `SELECT * FROM Bookmarks WHERE name LIKE ${search} AND folderId = ${folderId} ORDER BY \`${sortBy}\` ASC`;
+    var uid = req.session.uid;
+    if (search)
+    {
+        search = '% ' + req.query['search'] ? db.escape(req.query.search) : null + ' %';
+        sql    = `SELECT * FROM Bookmarks WHERE name LIKE ${search} AND folderId = ${folderId} AND uid = ${uid} ORDER BY ${sortBy} ASC`;
     }
-    else{
-        //sql = sqlSelect.from('Bookmarks').where({col:folderId=(${folderId}).order(${sortBy}).build();
-        sql = `SELECT * FROM Bookmarks WHERE folderId = ${folderId} ORDER BY ${sortBy} ASC`;
+    else
+    {
+        sql = `SELECT * FROM Bookmarks WHERE folderId = ${folderId} AND uid = ${uid} ORDER BY ${sortBy} ASC`;
     }
     console.log(sql);
     db.query(sql, function (err, bookmarks) {
         if (err)
         {
             console.error(err);
-            res.render('index', {error: err});
-            return;
+            req.reportedError = err;
         }
 
         var folders = getFolders(bookmarks);
-        var scope = {
+        var scope   = {
             bookmarks: bookmarks,
             showCreateDialog: req.showCreateDialog,
             showEditDialog: req.showEditDialog,
             showUploadDialog: req.showUploadDialog,
             showCreateFolderDialog: req.showCreateFolderDialog,
-            folders: folders
+            showConfirmDeleteDialog: req.showConfirmDeleteDialog,
+            showEditFolderDialog: req.showEditFolderDialog,
+            folders: folders,
+            error: req.reportedError
         };
 
-        if (scopeCallBack) {
+        if (scopeCallBack)
+        {
             scopeCallBack(scope);
         }
 
@@ -126,17 +103,9 @@ function getFolders(bookmarks) {
  * Selects information about passed in book and then
  * renders the delete confirmation page with the delete.ejs template
  */
-module.exports.confirmdelete = function (req, res) {
-    var id = req.params.book_id;
-    db.query('SELECT * from Bookmarks WHERE id =  ' + id, function (err, book) {
-        if (err)
-        {
-            reportedError = err;
-            res.redirect('/bookmarks');
-            return;
-        }
-        res.render('bookmarks/delete', {book: book[0]});
-    });
+module.exports.confirmDelete = function (req, res) {
+    req.showConfirmDeleteDialog = true;
+    renderEdit(req, res);
 };
 
 module.exports.add = function (req, res) {
@@ -162,33 +131,10 @@ module.exports.editBookmark = function (req, res) {
 };
 
 function renderEdit(req, res) {
-    console.info('List request', req.query);
-    var folderId = req.query['folderId'] ? db.escape(req.query.folderId) : 1;
-    var sortBy = req.query['sortBy'] ? req.query.sortBy : 'name';
     var id = req.query.id;
 
-    db.query(`SELECT * FROM Bookmarks WHERE folderId = ${folderId} ORDER BY ${sortBy}`, function (err, bookmarks) {
-        if (err) {
-            throw err;
-        }
-
-        var folders = getFolders(bookmarks);
-        console.log('folders ', folders);
-        console.log('id ', id);
-        var bookmarkItem = getBookmarkFromId(id, bookmarks);
-        console.log('Bm item ', bookmarkItem);
-
-        res.render('index', {
-            bookmarks: bookmarks,
-            showCreateDialog: req.showCreateDialog,
-            showEditDialog: req.showEditDialog,
-            folders: folders,
-            bookmarkItem: bookmarkItem
-        });
-    });
-
     function editDialogeScope(scope) {
-        scope.folders = getFolders(scope.bookmarks);
+        scope.folders      = getFolders(scope.bookmarks);
         scope.bookmarkItem = getBookmarkFromId(id, scope.bookmarks);
     }
 
@@ -212,6 +158,8 @@ function getBookmarkFromId(id, bookmarks) {
 module.exports.edit = function (req, res) {
     var id     = parseInt(req.params.bookId);
     var action = req.body.action;
+    var uid    = req.session.uid;
+
     delete req.body.action;
 
     var sql;
@@ -219,34 +167,36 @@ module.exports.edit = function (req, res) {
     if (action === 'Update')
     {
         var update = sqlQuery.update();
-        sql        = update.into('Bookmarks').set(req.body).where({id: id}).build();
+        sql        = update.into('Bookmarks').set(req.body).where({id: id, uid: uid}).build();
     }
     else
     {
-        sql = `DELETE FROM Bookmarks WHERE id=${db.escape(id)}`;
+        sql = `DELETE FROM Bookmarks WHERE id=${db.escape(id)} AND uid = ${uid}`;
     }
 
     console.log(action, sql);
     db.query(sql, function (err, response) {
         if (err)
         {
-            reportedError = err;
+            req.reportedError = err;
             console.error(err);
         }
         console.log(response);
-        res.redirect('/bookmarks');
+        renderIndex(req, res);
     });
 };
 
 module.exports.delete = function (req, res) {
-    var id = req.params.book_id;
-    db.query('DELETE from Bookmarks where id = ' + id, function (err) {
+    var id  = req.params.book_id;
+    var uid = req.session.id;
+
+    db.query(`DELETE from Bookmarks where uid = ${uid} AND id = ${id}`, function (err) {
         if (err)
         {
-            reportedError = err;
+            req.reportedError = err;
             console.error(err);
         }
-        res.redirect('/bookmarks');
+        renderIndex(req, res);
     });
 };
 
@@ -255,31 +205,39 @@ module.exports.delete = function (req, res) {
  * Does a redirect to the list page
  */
 module.exports.insert = function (req, res) {
-    var url         = db.escape(req.body.url);
-    var name        = db.escape(req.body.name);
-    var folderId    = db.escape(req.body.folderId);
-    var description = db.escape(req.body.description);
-    var keywords    = db.escape(req.body.keywords);
-    var favorite    = 0;
-    var folder      = "FALSE";
+    var sqlInsert           = sqlQuery.insert();
+    var newBookmark         = {};
+    newBookmark.url         = req.body.url;
+    newBookmark.name        = req.body.name;
+    newBookmark.folderId    = req.body.folderId;
+    newBookmark.description = req.body.description;
+    newBookmark.keywords    = req.body.keywords;
+    newBookmark.favorite    = 0;
+    newBookmark.folder      = "FALSE";
+    newBookmark.uid         = req.session.uid;
 
-  if (validUrl.isUri(url)){
-    console.log('Looks like an URI');
-  } else {
-    console.log('Not a URI');
-      res.render('403', { status: 403, bookmarks: bookmarks });
-  }
+    if (validUrl.isUri(newBookmark.url))
+    {
+        console.log('Looks like an URI');
+    }
+    else
+    {
+        console.log('Not a URI');
+        req.reportedError = {message: 'URI is inlaid', name: 'Bad Url given', status: 403};
+        renderIndex(req, res);
+        return;
+    }
 
-  var queryString = 'INSERT INTO Bookmarks (url, name, folderId, description, keywords, favorite, folder) VALUES (' + url + ', ' + name + ', ' + folderId + ', ' + description + ', ' + keywords + ', ' + favorite + ', ' + folder + ')';
-  console.log(queryString);
+    var queryString = sqlInsert.into('Bookmarks').set(newBookmark).build();
+    console.log(queryString);
 
     db.query(queryString, function (err) {
         if (err)
         {
-            reportedError = err;
+            req.reportedError = err;
             console.error(err);
         }
-        res.redirect('/bookmarks');
+        renderIndex(req, res);
     });
 };
 
@@ -304,39 +262,27 @@ function insertBookmark(bookmark, onError, done) {
 }
 
 module.exports.insertFolder = function (req, res) {
-    var url         = "NULL";
-    var folderId    = 1;
-    var keywords    = "NULL";
-    var name        = db.escape(req.body.name);
-    var description = db.escape(req.body.description);
-    var favorite    = 0;
-    var folder      = "TRUE";
-    var parent      = 1;
+    var sqlInsert         = sqlQuery.insert();
+    var newFolder         = {};
+    newFolder.folderId    = req.session.folderId ? req.session.folderId : 1;
+    newFolder.name        = req.body.name;
+    newFolder.description = req.body.description;
+    newFolder.favorite    = 0;
+    newFolder.folder      = true;
+    newFolder.uid         = req.session.uid;
 
-    var queryString = 'INSERT INTO Folders (name, parent) VALUES (' + name + ', ' + parent + ')';
+    var queryString = sqlInsert.into('Bookmarks').set(newFolder).build();
     console.log(queryString);
     db.query(queryString, function (err) {
         if (err)
         {
-            reportedError = err;
+            req.reportedError = err;
             console.error(err);
-            res.redirect('/bookmarks');
         }
-        //res.redirect('/bookmarks');
-    });
 
-    var queryString = 'INSERT INTO Bookmarks (url, name, folderId, description, keywords, favorite, folder) VALUES (' + url + ', ' + name + ', ' + folderId + ', ' + description + ', ' + keywords + ', ' + favorite + ', ' + folder + ')';
-    console.log(queryString);
-    db.query(queryString, function (err) {
-        if (err)
-        {
-            reportedError = err;
-            console.error(err);
-        }
-        res.redirect('/bookmarks');
+        renderIndex(req, res);
     });
 };
-
 
 /**
  * Updates a book in the database
@@ -348,59 +294,68 @@ module.exports.update = function (req, res) {
     var name        = db.escape(req.body.name);
     var description = db.escape(req.body.description);
     var keywords    = db.escape(req.body.keywords);
+    var uid         = req.session.uid;
 
-    if (validUrl.isUri(url)){
+    if (validUrl.isUri(url))
+    {
         console.log('Looks like an URI');
-    } else {
+    }
+    else
+    {
         console.log('Not a URI');
-        res.render('403', { status: 403, bookmarks: bookmarks });
+        req.reportedError = {message: 'URI is inlaid', name: 'Bad Url given', status: 403};
+        renderIndex(req, res);
     }
 
-    var queryString = 'UPDATE Bookmarks SET url = ' + url + ', name = ' + name + ', description = ' + description + ', keywords = ' + keywords + ' WHERE id = ' + id;
+    var queryString = 'UPDATE Bookmarks SET url = ' + url + ', name = ' + name + ', description = '
+        + description + ', keywords = ' + keywords + ' WHERE id = ' + id + ' AND uid = ' + uid;
     db.query(queryString, function (err) {
         if (err)
         {
-            reportedError = err;
+            req.reportedError = err;
             console.error(err);
         }
-        res.redirect('/bookmarks');
+        renderIndex(req, res);
     });
-
 
 };
 
 function search(req, res) {
+    var uid      = req.session.uid;
     var keywords = req.query['keywords'] ? db.escape(req.query.keywords) : 'keywords';
-    var sql = sqlSelect.from('Bookmarks').where({ col: sql.like(' keywords ') }).build();
+    var sql      = sqlSelect.from('Bookmarks').where({col: sql.like(' keywords '), uid: uid}).build();
 
     console.log(action, sql);
     db.query(sql, function (err, response) {
         if (err)
         {
-            res.redirect('/bookmarks');
-            throw err;
+            req.reportedError = err;
+            console.error(err);
         }
 
         console.log(response);
-        res.redirect('/bookmarks');
+        renderIndex(req, res);
     });
 }
 
-module.exports.search = function(req, res){
-  var search = req.body.keywords;
-  req.query.search = search;
-    renderIndex(req,res);
+module.exports.search = function (req, res) {
+    var search       = req.body.keywords;
+    req.query.search = search;
+    renderIndex(req, res);
 };
 
 module.exports.favorite = function (req, res) {
     var id          = req.query.id;
     var fav         = req.query.fav;
+    var uid         = req.session.uid;
+
     fav             = (fav + 1) % 2;
-    var queryString = 'UPDATE Bookmarks SET favorite = ' + fav + ' WHERE id = ' + id;
+    var queryString = 'UPDATE Bookmarks SET favorite = ' + fav + ' WHERE id = ' + id + ' AND uid = ' + uid;
     db.query(queryString, function (err) {
         if (err)
         {
-            throw err;
+            req.reportedError = err;
+            console.error(err);
         }
         renderIndex(req, res);
     });
@@ -416,7 +371,8 @@ module.exports.uploadFile = function (req, res, next) {
     upload(req, res, function (err) {
         if (err)
         {
-            return res.end("Error uploading file.");
+            req.reportedError = err;
+            console.error(err);
         }
         next();
     });
@@ -426,15 +382,18 @@ module.exports.defaultView = function (req, res) {
     renderIndex(req, res);
 };
 
-module.exports.sort = function(req, res){
-    var option = req.body.options;
+module.exports.sort = function (req, res) {
+    var option       = req.body.options;
     req.query.sortBy = option;
-    renderIndex(req,res);
-}
-
+    renderIndex(req, res);
+};
 
 module.exports.createFolder = function (req, res) {
-    console.log('making a foler');
     req.showCreateFolderDialog = true;
     renderIndex(req, res);
+};
+
+module.exports.showEditFolder = function (req, res) {
+    req.showEditFolderDialog = true;
+    renderEdit(req, res);
 };
